Correctly escape single quotes in the SQL query, and use INSTR to not interpret metacharacters.
This required fixing the escaping function to handle the case of the
replacement matching the source string.

Note: there may still be range errors in there! Not sure how to avoid
writing more than in the buffer, although at least I didn't succeed at
making it crash. It just seems to cut off the string, which in worst case
can cause a SQL error but should not be exploitable as long as there is
only one inserted parameter.

Much better approach would be adding support for bound parameters.